feat: Update user service to return GetUserResponse and add CreateUserRequest model
