+++ Logica/Models/CreateUserRequest.cs
@@ -1,0 +1,17 @@
+﻿using Data.Entities.Enums;
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace Logica.Models
+{
+    public class CreateUserRequest
+    {
+        public string Email { get; set; } = string.Empty;
+        public string Username { get; set; } = string.Empty;
+        public string Password { get; set; } = string.Empty;
+        public Role Role { get; set; } = Role.Shopper;
+    }
+}
