+++ Logica/Models/GetUserResponse.cs
@@ -1,0 +1,22 @@
+﻿using Data.Entities.Enums;
+using System;
+using System.Collections.Generic;
+using System.ComponentModel.DataAnnotations;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace Logica.Models
+{
+    public class GetUserResponse
+    {
+        public Guid Id { get; set; } = Guid.NewGuid();
+
+        public required string Email { get; set; } 
+
+
+        public required string  Username { get; set; }
+
+   
+    }
+}
